requirements.txt

google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.3.0.36
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
PyDrive==1.3.1
PyGetWindow==0.0.8
pylint==2.5.3
PyMsgBox==1.0.8
pyperclip==1.8.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQt5Designer==5.14.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==228
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.29

